Update patch set 6
Patch Set 6: Code-Review-1

There is something strange about exabgp in bagpipe-bgp. It's actually imported but it has not been added to requirements.txt and it's being installed with pip directly in tox.ini and devstack:

https://github.com/openstack/networking-bagpipe/blob/master/tools/tox_install.sh#L48
https://github.com/openstack/networking-bagpipe/blob/master/devstack/plugin.sh.bagpipe_bgp#L204

This may be some kind of temporary situation but it's making bagpipe-bgp not usable in RDO as-is. Could you clarify what's the status about exabgp?, if it's going to be kept as requirement, it should be added in requirements.txt (and pass through the new requirements process in upstream requirements project, i'd say), and it needs to be added in the RDO dependencies repo.

Patch-set: 6
Label: Code-Review=-1