Update patch set 7
Patch Set 7:

> @Thomas, for the directory change to work, you will need to add
 > `\neutron` to setup.cfg upstream in openstack bagpipe. I have the
 > change here staged in the spec file, but it fails while setuptools
 > is still using `/etc/bagpipe-bgp`

I would have expected it to be possible, in a spec file, to relocate a file (from where setuptools puts it in the rpm build root, to the final location on the filesystem when package unpacks).

If doable I would like to do this, and keep /etc/bagpipe-bgp as the default in networking-bagpipe to provide a smooth backward compat for people upgrading from a bagpipe-bgp installed from the github.com.

Patch-set: 7
Label: Code-Review=0