Drop requirements.txt as deps are handled manually
We are managing dependencies manually in spec so
we don't need requirements.txt as they can cause
conflicts.
networking-bgpvpn requires networking-bagpipe, to avoid
circular dependency we have already commented
networking-bgpvpn Requires: in spec but due to
auto dependency generator in EL9 the dependency
is automatically added, to get rid of it completely
let's remove requirements files as we manage
run time and build time dependency explicitly in
spec.

Change-Id: Id99c7f8e4b7a2ed03b4ca8ac6785bf30b1f5509a